addes token and created model e initialized middleware
