+++ .env.example
@@ -8,3 +8,6 @@ PORT_SEQU=""
 
 # MAIN
 PORT_MAIN=""
+
+# JWT
+PASSWORD_JWT=""
