+++ src/middleware/user.middleware.ts
@@ -1,0 +1,8 @@
+import { Injectable, NestMiddleware } from '@nestjs/common';
+
+@Injectable()
+export class UserMiddleware implements NestMiddleware {
+  use(req: any, res: any, next: () => void) {
+    next();
+  }
+}
