+++ src/service/login.service.ts
@@ -1,94 +1,9 @@
 import { Injectable } from "@nestjs/common";
 import { Error, Model } from "sequelize";
-import { INTEGER, ENUM, STRING, TEXT } from "sequelize";
-import { sequelize } from "src/sequelize/config.sequelize";
+import { Users, User, returnData, Departments} from "src/model/login.model";
 import * as bcrypt from 'bcrypt';
-
-// definição inicial da tabela de usuario, mas sera modificada e adicionada em um arquivo proprio
-const Users = sequelize.define('users', {
-  user_id: {
-    type: INTEGER,
-    autoIncrement: true,
-    primaryKey: true,
-    allowNull: false,
-  },
-  name: {
-    type: TEXT,
-    allowNull: false,
-  },
-  email: {
-    type: TEXT,
-    allowNull: false,
-  },
-  password: {
-    type: TEXT,
-    allowNull: false,
-  },
-  role: {
-    // type: ENUM('manager', 'analyst', 'admin'),
-    type: TEXT,
-    allowNull: true,
-  },
-  department_ID: {
-    type: INTEGER,
-    allowNull: true,
-  },
-}, {
-  timestamps: false,
-});
-
-const Departments = sequelize.define('departments', {
-  department_ID: {
-    type: INTEGER,
-    autoIncrement: true,
-    primaryKey: true,
-    allowNull: false,
-  },
-  department_name: {
-    type: TEXT,
-    allowNull: false,
-  },
-  description: {
-    type: TEXT,
-    allowNull: false,
-  },
-}, {
-  timestamps: false,
-});
-
-Departments.hasMany(Users, {
-  foreignKey: 'department_ID',
-  onDelete: 'CASCADE',
-  onUpdate: 'CASCADE',
-});
-
-Users.belongsTo(Departments, {
-  foreignKey: 'department_ID',
-  onDelete: 'CASCADE',
-  onUpdate: 'CASCADE',
-});
-
-class Department {
-  department_ID: number;
-  department_name: String;
-  description: String;
-}
-
-class User extends Department {
-  id_usuario: number;
-  name: String;
-  email: String;
-  password: String;
-  tipo_usuario: ['manager', 'analyst', 'admin'];
-  department_ID: number;
-  department: Department;
-}
-
-export class returnData {
-  status: number;
-  data?: Object;
-  message?: String;
-}
+import * as jwt from "jsonwebtoken";
+require('dotenv/config');
 
 @Injectable()
 export class loginService {
@@ -101,11 +16,9 @@ export class loginService {
       include: Departments,
     }).then((user: Model<User>) => {
 
+      // fiz essa funcao mas normalmente adiciono condicoes que retornam booleano diretamente no if
       function verifyPass(passVerify: string): boolean {
-        // const hash: string = bcrypt.hashSync(String(pass), 10);
         const value: boolean = bcrypt.compareSync(String(pass), passVerify);
-        console.log(value);
-        // console.log(hash);
         return value;
       }
 
@@ -113,18 +26,18 @@ export class loginService {
         throw "erro ao buscar usuario";
       } else if ((user.dataValues.email === email) && (verifyPass(String(user.dataValues.password))) === true) {
         dataUser.status = 200;
-        // aqui eu irei modificar para retornar o token
+        
         dataUser.data = {
           nome: user.dataValues.name,
           department: user.dataValues.department.department_name,
-          token: "token de acesso",
+          token: jwt.sign({nome: user.dataValues.name, department: user.dataValues.department.department_name}, process.env.PASSWORD_JWT, { algorithm: 'HS256' }),
         }
         
       } else {
         throw "usuario ou senha incorreto";
       }
     }).catch((err: Error) => {
-      // console.log(err);
+      console.log(err);
       if (err.name) {
         dataUser.status = 503;
         dataUser.message = err.name;
